+++ src/containers/review/LikeButtonContainer.jsx
@@ -1,4 +1,4 @@
-import { useUser } from "hooks/useUser";
+import { useUser } from "hooks/auth/useUser";
 import {
   useReviewLikeStatus,
   useToggleReviewLike,
